feat: implement friendboard nested comment
