+++ logic/friendboard/urls.py
@@ -6,9 +6,9 @@ from .views.posts import (
     FriendBoardPostDetailView,
 )
 from .views.posts.comments import FriendBoardPostsCommentsCreateView
+from .views.comments import FriendBoardCommentsDeleteView
 from .views.comments.comments import (
     FriendBoardCommentsCommentsCreateView,
-    FriendBoardCommentsCommentsDeleteView
 )
 
 # 술친구 urls.py
@@ -23,5 +23,5 @@ urlpatterns = [
     path('<int:pk>/', FriendBoardPostDetailView.as_view(), name="detail"),
     path('<int:pk>/comments', FriendBoardPostsCommentsCreateView.as_view(), name="createcomment"),
     path('comments/<int:pk>/comments', FriendBoardCommentsCommentsCreateView.as_view(), name="create_nested_comment"),
-    path('comments/<int:parent_pk>/comments/<int:pk>', FriendBoardCommentsCommentsDeleteView.as_view(), name="delete_nested_comment")
+    path('comments/<int:pk>', FriendBoardCommentsDeleteView.as_view(), name="delete_comment")
 ]
