+++ logic/friendboard/views/comments/comments.py
@@ -30,24 +30,3 @@ class FriendBoardCommentsCommentsCreateView(CreateView):
             return HttpResponseRedirect(self.get_success_url())
         else:
             raise Http404("Wrong Access")
-
-class FriendBoardCommentsCommentsDeleteView(DeleteView):
-    model = Comment
-
-    def get_parent_object(self):
-        return self.model.objects.get(pk=self.kwargs['parent_pk'])
-
-    def get_success_url(self):
-        return reverse_lazy('friendboard:detail', kwargs={'pk': self.get_parent_object().post.pk})
-
-    def delete(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        valpw = request.POST['valpw']
-
-        if self.object.password == valpw:
-            self.object.delete()
-            messages.info(request, '게시물 삭제에 성공했습니다.')
-        else:
-            messages.error(request, '패스워드가 다릅니다.')
-
-        return HttpResponseRedirect(self.get_success_url())
